requirements.txt
celery==3.1.23
tweepy==3.5.0
redis==2.10.5
click==6.6
neo4j-driver==1.0.2
progressbar2==3.10.1